move CoproductHint tests in CoproductHintSuite
