+++ core/src/test/scala/pureconfig/AnimalConfig.scala
@@ -1,0 +1,6 @@
+package pureconfig
+
+sealed trait AnimalConfig
+case class DogConfig(age: Int) extends AnimalConfig
+case class CatConfig(age: Int) extends AnimalConfig
+case class BirdConfig(canFly: Boolean) extends AnimalConfig
